Add new configuration and log handling setup
This commit introduces a new GitVersion.yml and nginx.conf configuration files, plus several shell scripts. In the code, it modifies the app's settings and adds a new logging mechanism. This tweak to logging is to emit a warning when a certain pattern is detected in the HTTP Request Header 'X-Original_URI', which is configured via 'requestWarningPattern' in the appsettings.json file.